// src/app/api-cinekeep/model/userListDetailsResponse.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PageResponseUserListItemResponse } from './pageResponseUserListItemResponse';
import { UserListResponse } from './userListResponse';


export interface UserListDetailsResponse { 
    list?: UserListResponse;
    appliedSortBy?: UserListDetailsResponse.AppliedSortByEnum;
    items?: PageResponseUserListItemResponse;
}
export namespace UserListDetailsResponse {
    export const AppliedSortByEnum = {
        OriginalOrderAsc: 'original_order.asc',
        OriginalOrderDesc: 'original_order.desc',
        TitleAsc: 'title.asc',
        TitleDesc: 'title.desc',
        ReleaseDateAsc: 'release_date.asc',
        ReleaseDateDesc: 'release_date.desc',
        VoteAverageAsc: 'vote_average.asc',
        VoteAverageDesc: 'vote_average.desc'
    } as const;
    export type AppliedSortByEnum = typeof AppliedSortByEnum[keyof typeof AppliedSortByEnum];
}