// src/app/api-cinekeep/model/userListItemResponse.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserListItemResponse { 
    id?: number;
    tmdbId?: number;
    mediaType?: UserListItemResponse.MediaTypeEnum;
    title?: string;
    posterPath?: string;
    backdropPath?: string;
    overview?: string;
    releaseDate?: string;
    voteAverage?: number;
    voteCount?: number;
    comment?: string;
    addedAt?: string;
}
export namespace UserListItemResponse {
    export const MediaTypeEnum = {
        Movie: 'movie',
        Tv: 'tv'
    } as const;
    export type MediaTypeEnum = typeof MediaTypeEnum[keyof typeof MediaTypeEnum];
}